chore: separate PostView to containers
